+++ ChangeLog.md
@@ -1,5 +1,8 @@
 # Revision history for use-c
 
-## 0.1.0.0  -- YYYY-mm-dd
+## 0.1.0.0  -- 2017-03-01
 
-* First version. Released on an unsuspecting world.
+* First version.
+* Scan struct definitions for a region attribute.  Unify regions from
+  declaration and the type of the first struct member.
+* Scan function declarations for pointers to a certain region.